requirements.txt
openai-agents
firecrawl
streamlit
firecrawl-py
requests
# Optional: for Hugging Face
transformers
torch
# Optional: for local models
ollama